pyproject.toml
[tool.poetry]
name = "streaming-demo"
version = "0.1.0"
description = ""
authors = ["Shinya Maeda <dosuken0122@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.343"
fastapi = "^0.104.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
